measurements: Harmonise temperature value type
...other measurements like Length or Mass are `f32` too.

+++ efb/src/measurements/constants.rs
@@ -15,7 +15,7 @@
 
 pub const FEET_IN_METER: f32 = 0.3048;
 pub const INCH_IN_METER: f32 = 0.0254;
-pub const KELVIN_IN_CELSIUS: f64 = 273.15;
+pub const KELVIN_IN_CELSIUS: f32 = 273.15;
 pub const METER_PER_SECONDS_IN_KNOTS: f32 = 1.943844;
 pub const NAUTICAL_MILE_IN_METER: f32 = 1852.0;
 pub const PI2: f32 = std::f32::consts::PI * 2.0;
